# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/users/email/{email}": {
      "get": {
        "operationId": "UsersController_findByEmail",
        "parameters": [
          {
            "name": "email",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    }
  },
  "info": {
    "title": "API Documentation",
    "description": "API documentation for NestJS - Mensageria application",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
...
}